requirements.txt
django<2.0
django-extensions
djangorestframework
markdown
django-filter<2
requests
Pillow
djangorestframework-jwt